Player is now created on game start and can be controlled by the user. Player can move right and left using the arrow keys. Player has an animated sprite. Background has a parralax effect based on player movement.
TODO:
Give player ability to jump and double jump.
Create camera class the passes on player velocity to any object when player is at max boarder.
Change boarders, limits and spawn locations based on screen size.
Clean up code base. Remove unneeded code and polish whats left.

+++ TheRedPicnic/Framework/background.h
@@ -42,9 +42,9 @@ protected:
 	Sprite* m_pMidgroundSprite;
 	Sprite* m_pGroundSprite;
 
-	int m_backgroundX[3];
-	int m_midgroundX[3];
-	int m_groundX[3];
+	float m_backgroundX[3];
+	float m_midgroundX[3];
+	float m_groundX[3];
 
 	float m_cameraVelocity;
